+++ test/unit/jasmine.json
@@ -1,12 +1,10 @@
 {
   "spec_dir": "test/unit",
   "spec_files": [
-    "test.js",
-    "util.js",
-    "module/module.js",
-    "module/module-collection.js"
+    "**/*.spec.js"
   ],
   "helpers": [
-    "../../node_modules/babel-register/lib/node.js"
+    "../../node_modules/babel-register/lib/node.js",
+    "setup.js"
   ]
 }
