+++ test/unit/module/module-collection.spec.js
