+++ test/unit/setup.js
@@ -1,0 +1,5 @@
+import 'babel-polyfill'
+import Vue from 'vue/dist/vue.common.js'
+import Vuex from '../../dist/vuex.js'
+
+Vue.use(Vuex)
